Move meta-files out of TLEs to util directory.
See #111

We'll do processing here to
* get tle files from Celestrak
* create a COMBINED file
* generate a SMD-specific file

We have to manualy maintail a list of SMD mission satellites
so that it can be updated and teh script generate new a SMD TLE based
on it.

The TLE files should be pulled and generated every night by cron.
It must be cautious not to overwrite files with damaged ones since
it's what the users' browsers will see.