Merge generation of SMD missions into getting from Celestrak
We now retrieve the file, save literal copies on disk.

We then process each to store an ALL tles and if the NORAD ID matches
on of ours from SMD, we store that in a SMD dict.

We finally output the deduped and sorted TLEs to ALL and SMD files.